Add support for setting desk name
This commit changes the following:
- Adds support for changing the desk name via `write_desk_name`
- Adds "n" and "w" to the demo to read and write the device name
- Renames the service/characteristic/cmd vars to reduce confusion
- Adds initial logging to the demo to make it clear it's not hanging

There exists a vendor characteristic 0xFE63 that supports writing the
desk name. The characteristic is Write+Notify, although the notification
appears to be an ack and doesn't contain the updated desk name. But the
update is observable by reading the standard (i.e., non-vendor-specific)
characteristic 0x2A00 that returns the BLE device name.

Note that the naming is confusing but deliberate: desk_name corresponds
to the vendor-specific desk_name, which is used by the BLE firmware to
derive the standard device_name. This is why I've called the functions
`write_desk_name` and `read_device_name`.

Further, I renamed the variables for the various services,
characteristics, and commands. Specifically, calling the commands
"uuids" was really confusing since they're just payload bytes in
characteristics. And the words std vs vendor now differentiate
between standard and vendor-specific services and characteristics.

I also renamed the height callback variables to make it more extensible,
as this PR introduces the ability to register callbacks for when the
desk_name is changed.